Fix problem in calculating the flesh score due to ints instead of floats

+++ src/document/Document.java
@@ -154,9 +154,9 @@ public abstract class Document {
 	{
 	    // TODO: You will play with this method in week 1, and 
 		// then implement it in week 2
-		int words = getNumWords();
-		int sentences = getNumSentences();
-		int syllables = getNumSyllables();
+		float words = getNumWords();
+		float sentences = getNumSentences();
+		float syllables = getNumSyllables();
 	    return 206.835 - (1.015 * (words / sentences)) - (84.6 * (syllables / words));
 	}
